+++ src/main/java/rapaio/data/stream/VSpots.java
@@ -27,7 +27,7 @@ package rapaio.data.stream;
 import rapaio.data.MappedVar;
 import rapaio.data.Mapping;
 import rapaio.data.Var;
-import rapaio.util.stream.StreamUtil;
+import rapaio.experiment.util.stream.StreamUtil;
 
 import java.io.Serializable;
 import java.util.Comparator;
