+++ src/main/java/rapaio/experiment/io/json/ml/JsonFeature.java
@@ -22,18 +22,32 @@
  *
  */
 
-package rapaio.io.json.stream;
+package rapaio.experiment.io.json.ml;
 
-import rapaio.io.json.tree.JsonValue;
+import rapaio.data.Var;
+import rapaio.experiment.io.json.tree.JsonValue;
 
-import java.io.Closeable;
-import java.io.IOException;
+import java.util.function.BiConsumer;
 
 /**
- * Created by <a href="mailto:padreati@yahoo.com">Aurelian Tutuianu</a> on 3/12/15.
+ * Created by <a href="mailto:padreati@yahoo.com">Aurelian Tutuianu</a> on 4/10/15.
  */
 @Deprecated
-public interface JsonInput extends Closeable {
+public class JsonFeature {
 
-    JsonValue read() throws IOException;
+    private final Var collector;
+    private final BiConsumer<JsonValue, Var> consumer;
+
+    public JsonFeature(Var collector, BiConsumer<JsonValue, Var> consumer) {
+        this.collector = collector;
+        this.consumer = consumer;
+    }
+
+    public void apply(JsonValue js) {
+        consumer.accept(js, collector);
+    }
+
+    public Var getResult() {
+        return collector;
+    }
 }
