+++ src/main/java/rapaio/experiment/io/json/stream/LzJsonAlgorithm.java
@@ -22,12 +22,12 @@
  *
  */
 
-package rapaio.io.json.stream;
+package rapaio.experiment.io.json.stream;
 
-import rapaio.io.json.stream.impl.IntEncoding;
-import rapaio.io.json.tree.JsonBool;
-import rapaio.io.json.tree.JsonNull;
-import rapaio.io.json.tree.JsonValue;
+import rapaio.experiment.io.json.stream.impl.IntEncoding;
+import rapaio.experiment.io.json.tree.JsonBool;
+import rapaio.experiment.io.json.tree.JsonNull;
+import rapaio.experiment.io.json.tree.JsonValue;
 
 /**
  * Created by <a href="mailto:padreati@yahoo.com">Aurelian Tutuianu</a> on 3/7/15.
