+++ src/main/java/rapaio/experiment/io/json/stream/LzJsonInput.java
@@ -22,9 +22,9 @@
  *
  */
 
-package rapaio.io.json.stream;
+package rapaio.experiment.io.json.stream;
 
-import rapaio.io.json.tree.*;
+import rapaio.experiment.io.json.tree.*;
 
 import java.io.DataInputStream;
 import java.io.IOException;
