+++ src/main/java/rapaio/experiment/io/json/tree/JsonObject.java
@@ -22,7 +22,7 @@
  *
  */
 
-package rapaio.io.json.tree;
+package rapaio.experiment.io.json.tree;
 
 import java.util.*;
 import java.util.stream.Stream;
