+++ src/main/java/rapaio/experiment/util/stream/FixedBatchSpliterator.java
@@ -22,7 +22,7 @@
  *
  */
 
-package rapaio.util.stream;
+package rapaio.experiment.util.stream;
 
 import java.util.Spliterator;
 import java.util.function.Consumer;
