+++ src/main/java/rapaio/experiment/util/stream/FixedBatchSpliteratorBase.java
@@ -22,7 +22,7 @@
  *
  */
 
-package rapaio.util.stream;
+package rapaio.experiment.util.stream;
 
 import java.util.Comparator;
 import java.util.Spliterator;
