+++ src/main/java/rapaio/experiment/util/stream/PartitioningSpliterator.java
@@ -22,7 +22,7 @@
  *
  */
 
-package rapaio.util.stream;
+package rapaio.experiment.util.stream;
 
 import java.util.LinkedList;
 import java.util.List;
